+++ backend/app/data_output_configuration/s3/schema.py
@@ -1,12 +1,12 @@
 from typing import Literal, Optional
 
+from app.configuration.environments.platform_service_configurations.schemas import (
+    AWSS3Config,
+)
 from app.data_output_configuration.base_schema import BaseDataOutputConfiguration
 from app.data_output_configuration.data_output_types import DataOutputTypes
 from app.data_output_configuration.s3.model import S3DataOutput as S3DataOutputModel
 from app.data_products.schema import DataProduct
-from app.environment_platform_service_configurations.schemas.s3_schema import (
-    AWSS3Config,
-)
 
 
 class S3DataOutput(BaseDataOutputConfiguration):