+++ backend/app/data_products/schema.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from app.data_product_types.schema import DataProductType
+from app.configuration.data_product_types.schema import DataProductType
 from app.data_products.status import DataProductStatus
 from app.shared.schema import ORMModel