+++ backend/app/datasets/schema_response.py
@@ -1,17 +1,17 @@
 from typing import Optional
 from uuid import UUID
 
+from app.configuration.data_product_lifecycles.schema import DataProductLifeCycle
+from app.configuration.data_product_settings.schema import DataProductSettingValue
+from app.configuration.domains.schema import Domain
+from app.configuration.tags.schema import Tag
 from app.data_outputs.schema import DataOutput as DataOutputBaseSchema
 from app.data_outputs_datasets.schema import DataOutputDatasetAssociation
-from app.data_product_lifecycles.schema import DataProductLifeCycle
-from app.data_product_settings.schema import DataProductSettingValue
 from app.data_products.schema import DataProduct
 from app.data_products_datasets.schema import DataProductDatasetAssociation
 from app.datasets.enums import DatasetAccessType
 from app.datasets.status import DatasetStatus
-from app.domains.schema import Domain
 from app.shared.schema import ORMModel
-from app.tags.schema import Tag
 
 
 class DataProductLink(DataProductDatasetAssociation):