+++ backend/tests/app/datasets/test_service.py
@@ -1,4 +1,11 @@
 from sqlalchemy.orm import selectinload
+
+from app.datasets.enums import DatasetAccessType
+from app.datasets.model import Dataset
+from app.datasets.service import DatasetService
+from app.roles import ADMIN_UUID
+from app.roles.schema import Prototype, Scope
+from app.settings import settings
 from tests import test_session
 from tests.factories import (
     DataProductDatasetAssociationFactory,
@@ -10,13 +17,6 @@ from tests.factories import (
     RoleFactory,
     UserFactory,
 )
-
-from app.datasets.enums import DatasetAccessType
-from app.datasets.model import Dataset
-from app.datasets.service import DatasetService
-from app.roles import ADMIN_UUID
-from app.roles.schema import Prototype, Scope
-from app.settings import settings
 
 
 class TestDatasetsService: