+++ backend/tests/app/events/test_router.py
@@ -1,13 +1,12 @@
+from app.core.authz import Action
+from app.roles.schema import Scope
+from app.settings import settings
 from tests.factories import DataOutputFactory, DataProductFactory, DatasetFactory
 from tests.factories.role import RoleFactory
 from tests.factories.role_assignment_data_product import (
     DataProductRoleAssignmentFactory,
 )
 from tests.factories.user import UserFactory
-
-from app.core.authz import Action
-from app.roles.schema import Scope
-from app.settings import settings
 
 ENDPOINT = "/api/events"
 
@@ -33,7 +32,7 @@ class TestEventsRouter:
             data_product_id=data_product.id,
         )
         response = client.put(
-            f"{"/api/data_products"}/{data_product.id}/about", json={"about": "active"}
+            f"{'/api/data_products'}/{data_product.id}/about", json={"about": "active"}
         )
         response = client.get(f"{ENDPOINT}/latest")
         assert response.json() is not None