+++ backend/tests/factories/data_output.py
@@ -1,11 +1,11 @@
 import factory
+
+from app.data_outputs.model import DataOutput
+from app.data_outputs.status import DataOutputStatus
 from tests import test_session
 from tests.factories.data_product import DataProductFactory
 from tests.factories.platform_service import PlatformServiceFactory
 from tests.factories.s3_data_output import S3DataOutputFactory
-
-from app.data_outputs.model import DataOutput
-from app.data_outputs.status import DataOutputStatus
 
 
 class DataOutputFactory(factory.alchemy.SQLAlchemyModelFactory):